Deprecate the 'size' attribute of track nodes
In the v2 CLI version of the mapwarper, this attibute told the
pixelsize of the image the node was created in. The idea was that the
smoother might use it to offset the track no more than this distance
in order to get a less wiggly line. This turned out never to be
necessary (nor was it ever clear _how_ this would work anytway).

For the GUI warper I never bothered to figure out what this number
_should_ mean (the pixelsize of the map tiles used, or the UI pixels
that determine the precision of the mouse click location?) since it
wasn't used for anything, so the GUI version always just put "1" as
a placeholder.

Since the v2 .vect parser is now deleted and I'm free to adapt the
parser, switch to not even emitting the "1" for nodes created
henceforth. Existing nodes with sizes (even ones that say "1") keep
their sizes such as to reduce diff sizes.